minor  changes
