+++ src/dataDisplay/index.ts
@@ -8,4 +8,3 @@ export { default as Layout } from './Layout';
 export { default as Text } from './Text';
 export { default as Title } from './Title';
 export { default as Section } from './Section';
-export { default as Identicon } from './Identicon';
