+++ src/utils/index.ts
@@ -1,1 +1,2 @@
+export { default as Identicon } from './Identicon';
 export * from './modals'
